--- a/server/Cargo.toml
+++ b/server/Cargo.toml
@@ -24,6 +24,7 @@ byte-unit = { version = "5.1.2", default-features = false, features = [
 bytes = "1.5.0"
 chrono = "0.4.31"
 clap = { version = "4.4.13", features = ["derive"] }
+dashmap = "5.5.3"
 figlet-rs = "0.1.5"
 figment = { version = "0.10.13", features = ["json", "toml", "env"] }
 flume = "0.11.0"