--- a/server/src/streaming/partitions/messages.rs
+++ b/server/src/streaming/partitions/messages.rs
@@ -135,11 +135,9 @@ impl Partition {
         count: u32,
     ) -> Result<Vec<Arc<Message>>, Error> {
         let (consumer_offsets, consumer_id) = match consumer {
-            PollingConsumer::Consumer(consumer_id, _) => {
-                (self.consumer_offsets.read().await, consumer_id)
-            }
+            PollingConsumer::Consumer(consumer_id, _) => (&self.consumer_offsets, consumer_id),
             PollingConsumer::ConsumerGroup(consumer_group_id, _) => {
-                (self.consumer_group_offsets.read().await, consumer_group_id)
+                (&self.consumer_group_offsets, consumer_group_id)
             }
         };