Resolve https://github.com/iggy-rs/iggy/issues/226

I used the `DashMap` to replace the original `RwLock<HashMap>`, and because the locking of `DashMap` is synchronous semantics, I had to use clone in the `fn store_offset` to avoid the lock crossing await.