@hubcio 2024-01-10
Could you please run `bench`?
cargo run --release --bin iggy-bench send tcp
cargo run --release --bin iggy-bench poll tcp
cargo run --release --bin iggy-bench send-and-poll tcp

do it on before/after and post results here