@spetz 2024-01-10
Looks good, have you had a chance to run the benchmarks just to compare the results? I guess that there should be not much of a difference :)