@YHM404 2024-01-10
@hubcio @spetz I run these benchs on Mac m1. It seems no significant difference.

### Before:

**cargo run --release --bin iggy-bench send tcp:**
```
total throughput: 991.80 MB/s, 991800 messages/s, average latency: 10.08 ms, average  throughput: 99.18 MB/s, total duration: 10.08 s
```
**cargo run --release --bin iggy-bench poll tcp:**
```
total throughput: 2192.55 MB/s, 2098130 messages/s, average latency: 4.68 ms, average  throughput: 219.25 MB/s, total duration: 4.77 s
```

**cargo run --release --bin iggy-bench send-and-poll tcp:**
```
2024-01-10T16:26:49.491908Z  INFO iggy_bench::benchmarks::send_and_poll_benchmark: Benchmark: send and poll messages, transport: tcp, total messages: 10.00 M, total size: 9.3 GiB bytes, 10 streams, 1000 messages per batch, 1000 batches, 1000 bytes per message, 10 producers, 10 consumers

2024-01-10T16:26:49.491921Z  INFO iggy_bench::benchmark_runner: Producer results: total throughput: 1280.45 MB/s, 1280447 messages/s, average latency: 7.81 ms, average throughput: 64.02 MB/s, total duration: 7.81 s, Consumer results: total throughput: 1271.08 MB/s, 1216344 messages/s, average latency: 8.15 ms, average  throughput: 63.55 MB/s, total duration: 8.22 s
2024-01-10T16:26:49.491924Z  INFO iggy_bench::benchmark_runner: Results: total throughput: 1275.64 MB/s, 1247573 messages/s, average latency: 15.96 ms, average  throughput: 63.78 MB/s, total duration: 16.03 s
```

### After:

**cargo run --release --bin iggy-bench send tcp:**

```
total throughput: 1053.67 MB/s, 1053673 messages/s, average latency: 9.49 ms, average  throughput: 105.37 MB/s, total duration: 9.49 s
```

**cargo run --release --bin iggy-bench poll tcp:**

```
total throughput: 2024.59 MB/s, 1937406 messages/s, average latency: 5.06 ms, average  throughput: 202.46 MB/s, total duration: 5.16 s
```

**cargo run --release --bin iggy-bench send-and-poll tcp:**
```
2024-01-10T16:40:40.952999Z  INFO iggy_bench::benchmark_runner: Producer results: total throughput: 1402.16 MB/s, 1402158 messages/s, average latency: 7.13 ms, average throughput: 70.11 MB/s, total duration: 7.13 s, Consumer results: total throughput: 1391.61 MB/s, 1331687 messages/s, average latency: 7.44 ms, average  throughput: 69.58 MB/s, total duration: 7.51 s
2024-01-10T16:40:40.953002Z  INFO iggy_bench::benchmark_runner: Results: total throughput: 1396.75 MB/s, 1366014 messages/s, average latency: 14.57 ms, average  throughput: 69.84 MB/s, total duration: 14.64 s
```